Move device-related code into its own subdirectory
