+++ pyquil/device/tests/test_device.py
